HST colour transformations now read from table
Tables are generated using the modelling scripts in `scripts/` and data
tables are stored in `androphotsys/data`. Use the
`query_colour_transformation` function to get the transformation
coefficients.

This system means less degree for error in hard coding coefficients, and
it also becomes easier to print out a table of these coefficients too
for publication.